ECTS-Punkte von int auf double umgestellt
- Datenmodell, ViewModels, DTOs und Validierung auf double für ECTS angepasst
- Neue Migration für ECTS als double, alte Migration entfernt
- Eingabeprüfung für Dezimaltrennzeichen (Komma/Punkt) erweitert
- Standardkultur auf "de-DE" gesetzt (Zahlenformat)
- Beispielmodule mit ECTS-Werten wie 7.5/2.5 in DB-Init
- Benachrichtigungen enthalten jetzt Benutzerrolle
- Projektdateien und Migrationen aktualisiert

+++ Modulverwaltungssoftware/EditWithCommentsView.xaml.cs
@@ -241,7 +241,7 @@ namespace Modulverwaltungssoftware
             ResetValidationHighlights();
             
             // ✅ Basis-Validierung
-            if (!ValidateBasicInputs(out int ects, out int workloadPraesenz, out int workloadSelbststudium))
+            if (!ValidateBasicInputs(out double ects, out int workloadPraesenz, out int workloadSelbststudium))
             {
                 return; // Fehlermeldung wurde bereits angezeigt
             }
@@ -374,7 +374,7 @@ namespace Modulverwaltungssoftware
         /// <summary>
         /// ✅ NEU: Validiert Basis-Eingaben und hebt fehlerhafte Felder hervor
         /// </summary>
-        private bool ValidateBasicInputs(out int ects, out int workloadPraesenz, out int workloadSelbststudium)
+        private bool ValidateBasicInputs(out double ects, out int workloadPraesenz, out int workloadSelbststudium)
         {
             ects = 0;
             workloadPraesenz = 0;
@@ -389,7 +389,7 @@ namespace Modulverwaltungssoftware
             }
 
             // ECTS prüfen
-            if (!int.TryParse(EctsTextBox.Text, out ects) || ects <= 0)
+            if (!double.TryParse(EctsTextBox.Text, out ects) || ects <= 0)
             {
                 ValidationHelper.MarkAsInvalid(EctsTextBox, "Bitte gültige ECTS-Punktzahl (> 0) eingeben.");
                 isValid = false;
